Allow label selectors for snapshot policy
The SnapshotPolicy CRD supported single way of defining the PVCs it will
create periodic snapshot of. This meant for instance that increasing
a replica count on a StatefulSet required a knowledge of
SnapshotPolicies and manual interventioni there as well, otherwise the app
would end up with insufficient count of snapshots.

This allows specifying a list of label selectors matching an app to make
this less error prone.